+++ flake.nix
@@ -1,8 +1,7 @@
 {
-  description = "tailwind's description";
+  description = "tailwind-haskell";
   inputs = {
-    nixpkgs.url = "github:nixos/nixpkgs/568e0bc498ee51fdd88e1e94089de05f2fdbd18b";
-    tailwind-nix.url = "github:srid/tailwind-nix";
+    nixpkgs.url = "github:nixos/nixpkgs/554d2d8aa25b6e583575459c297ec23750adb6cb";
     flake-utils.url = "github:numtide/flake-utils";
     flake-compat = {
       url = "github:edolstra/flake-compat";
@@ -55,7 +54,7 @@
                   haskell-language-server
                   ormolu
                   pkgs.nixpkgs-fmt
-                  inputs.tailwind-nix.defaultPackage.${system}
+                  pkgs.nodePackages.tailwindcss
                 ]);
           };
       in