+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `--output`
 - Fix quoted source paths breaking tailwind.config.js syntax
 - Workaround Tailwind not crashing with non-zero exitcode
+- [Nix] Switch to nixpkgs' tailwindcss.
 
 ## 0.1.0.0 -- 2022-01-08